Create README.md
Looking to make new friends and people interested in working on a crypto project ( I have a great Coin Idea and im using my experience from working for the top 3 banks and believe we can make something great from this). I have a concept that will be secure, require much less energy form miners, concept has been a project for months just need to get it going on the computer. What I lack in coding I make up with business and financial word knowledge. By The way I have a great website Idea that will blow Zillow out of the water and it has an incentive plan that is amazing and the legal side I have already had my attorney check it out and we are good to go.